+++ packages/web/src/lib/view/parts-list/state/state-serializer.ts
@@ -8,20 +8,21 @@ import { Result } from '@praha/byethrow'
 import {
   compressToUrlSafeString,
   decompressFromUrlSafeString,
-} from './filter/compression'
-import { type Filter } from './filter/filters-core'
+} from '../filter/compression'
+import { type Filter } from '../filter/filters-core'
 import {
   normalizeSlotKey,
   parseFilter,
   type FiltersPerSlot,
-} from './filter/serialization'
+} from '../filter/serialization'
+import { parseSort } from '../sort/sort'
+import type { SortKey, SortOrder } from '../sort/sort'
+
 import { type DeserializeError, VALID_SLOTS } from './shared'
 import { toSlotParamValue } from './slot-utils'
-import type { SortKey, SortOrder } from './sort'
-import { parseSort } from './sort'
 
 // Filterは filters.ts からエクスポート
-export type { Filter } from './filter/filters-core'
+export type { Filter } from '../filter/filters-core'
 
 /**
  * URL共有用の状態（スロット、フィルタ、並び替え）
